Improve util function to accept rootValue
Prior to this commit it was not possible to use `runAuthenticatedQuery`
with a `rootValue`, instead you'd use `runQuery` and provide the
authenticated values. Whaa, whaa.

Before considering this change I audited the usage of
`runAuthenticatedQuery` and found that in only one place were we not
taking the default value for userID, and even in that one place we
actually WERE using the default value! So, I floated the idea on Slack
and it sounded ok to @alloy.

From there I simply copied the signature of `runQuery` and scooped out
the defaults to override them in the body of the function. Since you
can't mutate a function param without being yelled at, I copied the
`rootValue` into a local const and set the appropriate values.

Technically, there is a regression here where you cannot use an
incorrect userID, but since that "feature" wasn't being used, I think
it's ok - open to other thoughts on this though!!