Take in loaders in more places
I'm not sure why you can always seem to take in these four arguments,
but per feedback from @alloy on #707, I was doing this wrong. My
assumption was that there was something special about the `resolve`
functions I was working with that allowed it to take in these four, but
maybe you can just always do this?

Anyway, this should be good now!