Revert even more loader work
During pair swap, I did some incorrect work in #707 that I thought was
safely reverted in #714, but I was wrong. There were still other places
where we had broken functions and this was only caught after deployment
to production.

This commit should revert the rest of the broken changes from #707 and
fix the bugs seen in production.