[Artist] Guard against empty ID string.
This specific guard is needed because from Emission we conditionally
fetch a ‘selected artist’ for the Works for You view, but using an extra
boolean for this proved problematic because it meant that the ID
variable would be unused and GraphQL would complain about that. Instead
we always send the‘selected artist’ request, but with an empty string
when we don’t actually want it.