[dev] Make sure app is fully started on initial request.
Previously the app would start launching when the initial request comes
in, but not yet be finished launching, which would lead to the inital
request failing. This was especially painful when that initial request
was an introspection query being performed by e.g. GraphiQL and then
leading to the app not knowing the schema.