[rateLimiter] Trust clients based on having set a DataDog header.
Because the DD libs already patch all outgoing HTTP requests to include
this, meaning we don’t need to go through all our code to find the
places where we make requests and update each of those.

Additionally we shouldn’t rely on these types of headers at all in the
long run, so using this for pragmatic reasons seems fine.