Return null shipping info if artwork is not acquirable
Pages for artworks which were previously enrolled in an e-commerce
program still display shipping information provided during enrollment in
the program.

This creates an inconsistent user experience. If an artwork is removed
from a program, we should not display any shipping information
previously entered during enrollment.

This commit updates Metaphysics to return a `null` value for
`shippingInfo` unless the artwork is `acquirable`, designating active
enrollment in one or more of our e-commerce programs.