Update stale expectation
The format of the date for this show comes from a helper in
`src/lib/date.ts` called `exhibitionPeriod` and it has logic about
showing or not the year. That behavior is tested over in the helper so
covering it here isn't necessary so I just updated the expectation to
include this concept.