Partially implements the check to avoid gravity query for analytics
When the query is not checking for anything other than analytics MP
should not query gravity. This PR is partially implements that but the
parseFieldASTsIntoArray method currently doesn't go deeper into
feildASTs and in this case the field is within an InlineFragment
(because of ..... on Partner { _id } in vortex stitching file) so this
really doesn't work. Will follow up to fix this.