Display when a sale is available for live bidding, when available
There are 5 possible states a live sale could be in (live as in live
auction integration):

1. Published on Artsy, open for registration, but not yet open for any
   kind of bidding
2. Open for registration & pre-bidding, but not yet open for live
   bidding
3. Registration closed, pre-bidding open to those who previously
   registered, but not yet open for live bidding
4. Open for live bidding
5. Closed

_Thanks to Melanie Borinstein for this context!_

For live auctions, we want to display the date an auction becomes
available for live bidding (4), not when the auction enters any previous
state (1-3). We currently display the date corresponding to state 2,
when an auction is open for registration and pre-bidding.

ticket: https://artsyproduct.atlassian.net/browse/DISCO-944 :lock: