Introduce Sale.userNeedsIdentityVerification field
Problem:

Logic about whether the current user needs to undergo identity
verification is scattered around Reaction and Eigen, increasing risk
that business logic starts to drift and generally is a larger code
surface area than needed.

Solution:

Introduce new field to `Sale` which is the source of truth as to whether
the current user needs to undergo identity verification or not for the
sale in question.

Why introduce this field to `Sale` instead of `me`, `User`, or `Bidder`?

- See this [Slack convo][0]

- Because there are times where the user isn't yet registered to bid, but we
need to know if the current user needs to undergo identity verification
or not, so extending `Bidder` doesn't seem to make sense.

- Moreover, there are even times whether there's no reference to a
current user (`me`). In that case, we want to display that the "current
user" needs to undergo identity verification, so extending `me`/`User`
doesn't seems to make sense.

- We always have reference to a `Sale`, so extending that type with this
field makes the most sense. This is a similar to
`Sale.registrationStatus`.

Jira: https://artsyproduct.atlassian.net/browse/AUCT-983

---

* Rename requiresIdentityVerificationFor ->
userNeedsIdentityVerification to clarify that this is a field for the
user and that it is about whether the user needs to do something, not
about whether the user requires something

[0]:https://artsy.slack.com/archives/CA3D02SDN/p1587581641465800