Datadog Route Tracing
Problem

Currently, the V1 and V2 routes are being as originating from the
URL in Datadog. This is preventing us from being able to track how
many requests are hitting each endpoint and determine how many
clients are still using the V1 routes.

Solution

This change initializes Datadog earlier so that we will be able to
differentiate the V1 and V2 APIs.

Due to the way that Datadog internally recognizes "valid" routes, if
dd-trace has not loaded and monkey patched express prior to a
route being defined. Datadog will associated it with the highest
applicable bucket. In this case that would be `/`.

Additionally, there is a minor refactoring that removes a extra
express app that wasn't required.