Fix total count for v2 artworks resolver
The field is currently only used with `ids` param passed in, and in that
case, Gravity /artworks endpoint will not paginate and, instead return
all matching artworks in the response.

We should calculate total count based on the response, instead of the
`ids` param, which could include arbitrary/invalid values.

https://github.com/artsy/gravity/blob/6df1e3524d1bcb60a6c6773a487b4a560788b5db/app/api/v1/artworks_endpoint.rb#L166-L169