Enable Datadog GraphQL Tracing
Problem

Datadog tracing is not enabled for graphql and enabling the graphql plugin doesn't "just work" due to a bug in `dd-trace-js`.

For more information read the following issue:
https://github.com/DataDog/dd-trace-js/issues/951

Solution

The solution is to patch our copy of Datadog with the intended fix and upgrade once the bug has been resolved.