feat: allow clients to fetch a secured public API key for Algolia queries
Allow clients (like Force and Eigen) to retrieve a secured API key to search
against our Algolia backend.

If a user is authenticated, the userToken will be the userID and any rate limit
options set on the parent search API key will apply to the user.

If a user is unauthenticated, the userToken will be blank and any rate limit
options set on the parent search API key will apply to the client's IP address.

We might consider moving the generation of these signed API keys to Gravity if
we anticipate needing different controls based on the client application. For
example, we might want to specify a list of referrers for tokens used on web
clients.

**Request**

```graphql
{
  algolia {
    apiKey
  }
}
```

**Response**

```json
{
  "data": {
    "algolia": {
      "apiKey": "<token>"
    }
  }
}
```