Install yarn conditionally
I updated the general setup script to install yarn because I think it's
easier to do as much of that generalized setup as possible in the same
place.

The change here includes a fall-back to install yarn if it hasn't
already been installed specifically for individuals who might come to
this repository without following the generalized setup script in
potential, i.e. external contributors.

Secondarily, I don't think we should force anyone who runs this script
to update to the latest version of yarn if the version they have will
function.