feat: stitch Gravity-backed algolia type into system root query field (#3374)
* chore: remove previous algolia implementation

* feat: stitch Gravity-backed algolia type into system root query field

Note: In the example below, we're using `__typename` to trigger the full query
resolution. Without it, `system` comes back `null`. I've seen this behavior for
other queries that only rely on stitched fields. It seems like you need some
part of your query that is not relying on stitched resolvers for the full data
resolution to trigger.

**Request**

```graphql
{
  system {
    __typename
    algolia {
      apiKey
      appID
      indices {
        name
        displayName
      }
    }
  }
}
```

**Response**

```json
{
  "data": {
    "system": {
      "__typename": "System",
      "algolia": {
        "apiKey": "<api key>",
        "appID": "<app id>",
        "indices": [
          {
            "name": "Artist_staging",
            "displayName": "Artist"
          }
        ]
      }
    }
  }
}
```

* chore: update stitched gravity schema snapshot