feat: add trim option to strip away lot label past first alphanumeric chunk
We've encountered cases where sale artworks in Gravity have non-numeric
characters to store metadata related to the lot. This isn't a recommended use
of the lot label field. As an alternative, we've recommended that colleagues
use the `blurb` field which is also exposed on the same surfaces across web and
app ([Slack](https://artsy.slack.com/archives/C02RAMW9QQJ/p1664364783118159)).

This commit updates the lotLabel resolver to accept an optional `trim` argument
(default=false) to strip away anything past the first alphanumeric chunk from
the Gravity-backed input. We'll use this in clients that need a shorter display
string but we might consider adding this formatting and perhaps validation
upstream in Gravity.