feat: specify a custom component for the new section
In this case we override the default artworks section presentation
with a component that is defined on the client side.

See the Eigen branch `onyx/home-view-ks-demo-custom-component` for the
corresponding client-side implementation.

(If it doesn't exist on the client -- e.g on an old Eigen version --
then the client will sensibly fall back to the default artworks section
implementation.)