chore: remove support for HomeViewSectionGalleries
This section was a misnomer, since it never served up a
connection of galleries.

The one use of it has been replaced with HomeViewSectionCard.

Which means we can now remove this section type, thus freeing
up that name in the future for a new section type that actually does
return galleries.