feat(home view): add a HomeViewSectionCard (singular) section type (#6116)
* feat: create a new singular HomeViewSectionCard section type

* feat: convert GalleriesNearYou to the new HomeViewSectionCard type

* feat: add support for additional button/cta text

* chore: remove support for HomeViewSectionGalleries

This section was a misnomer, since it never served up a
connection of galleries.

The one use of it has been replaced with HomeViewSectionCard.

Which means we can now remove this section type, thus freeing
up that name in the future for a new section type that actually does
return galleries.

* refactor: refer to component attrs when returning card

* chore: expose a ts type for HomeViewCard

* refactor: avoid non-canonical web url