feat(home view): allow section declaration to specify a minimum required Eigen version (#6297)
* chore(deps): update node types to match current 18.15.0

In order to use Array.at() in the next commit, I need to update our
Node types to match the actual version of Node we are using.

I also had to tweak one file to conform to the updated types.

* feat: add a utility for parsing Eigen version numbers

* feat: add a predicate that asserts a minimum version number

* feat: allow section declaration to specify a minimum Eigen version

* feat: enforce the minimum Eigen version

But only if there is an actual Eigen version to consider.

Otherwise we are presumably coming from a different client
altogether, and would not want to gate this section.

* refactor: fix filename casing

* test: fix desktop UA string

* refactor: assume UA is string, not array

* test: add an Android case for good measure