fix: param/body construction on createNewGeminiAssetLoader
We were passing acl as an query param on the URL path but also providing acl in
the request body.

This commit is focused on fixing this one instance but it's possible that an
[earlier PR][#7002] to fix array serialization in mutations and POST request
bodies introduced incompatibility with combining both query params in the URL
path and a request body at the same time.

[#7002]: https://github.com/artsy/metaphysics/pull/7002